Docs: next-auth.d.ts 파일 추가
